+++ client/preload_invite.js
@@ -1,0 +1,8 @@
+const { contextBridge, ipcRenderer } = require("electron");
+
+contextBridge.exposeInMainWorld("invite", {
+  submit: (payload) => ipcRenderer.send("invite-submit", payload),
+  cancel: () => ipcRenderer.send("invite-cancel"),
+  onError: (cb) => ipcRenderer.on("invite-error", (_e, data) => cb(data)),
+});
+